feat: agregar gestión de profesores; implementar servicio y actualizar formularios de usuario y curso

+++ src/app/components/admin/users/user-admin.component.ts
@@ -13,7 +13,7 @@ export class UserAdminComponent implements OnInit {
   users: any[] = [];
   isLoading: boolean = true;
 
-  constructor(private authService: AuthService, private router: Router) {}
+  constructor(private authService: AuthService) {}
 
   ngOnInit(): void {
     this.loadUsers();
@@ -33,22 +33,34 @@ export class UserAdminComponent implements OnInit {
     );
   }
 
-  deleteUser(userId: number): void {
-    this.authService.deleteUser(userId).subscribe(
-      (response: any) => {
-         Swal.fire({
-           title: 'Usuario eliminado correctamente',
-           icon: 'success',
-         });
-
-        this.loadUsers();
-      },
-      (error) => {
-        Swal.fire({
-          title: 'Error al eliminar usuario',
-          icon: 'error',
-        });
+  deleteUser(userId: number, userName: string, userLastname: string): void {
+    Swal.fire({
+      title: `¿Estás seguro que quieres borrar a ${userName}?`,
+      text: '¡Los cambios no se pueden modificar!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#4b6584',
+      cancelButtonColor: '#c85a42',
+      confirmButtonText: 'Sí, borrar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.authService.deleteUser(userId).subscribe(
+          (response: any) => {
+            Swal.fire({
+              title: '¡Borrado!',
+              text: `El usuario ${userName} ${userLastname} ha sido borrado.`,
+              icon: 'success',
+            });
+            this.loadUsers();
+          },
+          (error) => {
+            Swal.fire({
+              title: 'Error al eliminar usuario',
+              icon: 'error',
+            });
+          }
+        );
       }
-    );
+    });
   }
 }
